moved heapify functions to a separate file
